refactor(llm-provider): 优化数据转换和统一计数类型
- 使用Stream流式处理代替传统for循环，提升代码简洁性和可读性
- 引入Objects::nonNull过滤空元素，避免潜在空指针异常
- 将模型计数相关字段类型从Long改为Integer，确保一致性
- 删除LlmProviderDetailVo中不必要的apiKey属性，提高安全性
- 为LlmProviderListDto新增完善的字段注释，提升代码可维护性
- 统一修改多处测试代码以匹配类型变更

+++ medicine-admin/src/main/java/cn/zhangchuangla/medicine/admin/model/dto/LlmProviderListDto.java
@@ -16,31 +16,73 @@ import java.util.Date;
 @AllArgsConstructor
 public class LlmProviderListDto {
 
+    /**
+     * 主键ID。
+     */
     private Long id;
 
+    /**
+     * 提供商名称。
+     */
     private String providerName;
 
+    /**
+     * 基础请求地址。
+     */
     private String baseUrl;
 
+    /**
+     * 提供商描述。
+     */
     private String description;
 
+    /**
+     * 状态。
+     */
     private Integer status;
 
+    /**
+     * 排序值。
+     */
     private Integer sort;
 
+    /**
+     * 创建人。
+     */
     private String createBy;
 
+    /**
+     * 更新人。
+     */
     private String updateBy;
 
+    /**
+     * 创建时间。
+     */
     private Date createdAt;
 
+    /**
+     * 更新时间。
+     */
     private Date updatedAt;
 
-    private Long modelCount;
+    /**
+     * 模型总数。
+     */
+    private Integer modelCount;
 
-    private Long chatModelCount;
+    /**
+     * 对话模型数量。
+     */
+    private Integer chatModelCount;
 
-    private Long rerankModelCount;
+    /**
+     * 重排模型数量。
+     */
+    private Integer rerankModelCount;
 
-    private Long embeddingModelCount;
+    /**
+     * 向量模型数量。
+     */
+    private Integer embeddingModelCount;
 }
